requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
cloudinary==1.38.0
django-cloudinary-storage==0.0.12
